slightly refactored binarized integration test components
