+++ test/model_generation/integration_test/test_single_op_models/test_binarized/__init__.py
@@ -7,7 +7,6 @@ import tensorflow as tf
 import larq_compute_engine as lce
 from typing import Optional, Tuple, Type, Any, Union, NamedTuple
 
-from tflite2xcore.utils import get_bitpacked_shape, unpack_bits  # type: ignore # TODO: fix this
 from tflite2xcore.xcore_schema import (  # type: ignore # TODO: fix this
     Tensor,
     ExternalOpCodes,
@@ -27,12 +26,12 @@ from tflite2xcore.converter import CleanupManager  # type: ignore # TODO: fix th
 from tflite2xcore.model_generation import Configuration, TFLiteModel, Hook
 from tflite2xcore.model_generation.evaluators import LarqEvaluator
 from tflite2xcore.model_generation.runners import Runner
-from tflite2xcore.model_generation.converters import KerasModelConverter
+from tflite2xcore.model_generation.converters import KerasModelConverter, LarqConverter
 from tflite2xcore.model_generation.data_factories import InputInitializerDataFactory
 from tflite2xcore.model_generation.utils import parse_init_config
 
 from .. import (
-    IntegrationTestRunner,
+    BinarizedTestRunner,
     test_reference_model_regression,
     test_converted_single_op_model,
     test_mean_abs_diffs,
@@ -144,8 +143,12 @@ class RemoveSingleOutputOperatorPass(OutputTensorMatchingPass):  # type: ignore 
         self._done = True
 
 
-class LarqConverter(KerasModelConverter):
-    """ Converts a Larq model to a TFLite model. """
+class LarqSingleOpConverter(LarqConverter):
+    """ Converts a larq composite TFL model to a single op TFL model.
+    
+        This converter is to work around the fact that some larq ops
+        cannot be directly generated from keras layers.
+    """
 
     def __init__(
         self,
@@ -159,12 +162,7 @@ class LarqConverter(KerasModelConverter):
         self._remove_last_op = remove_last_op
 
     def convert(self) -> None:
-        self._model = lce.convert_keras_model(
-            self._input_model_hook(),
-            inference_input_type=tf.int8,
-            inference_output_type=tf.int8,
-            experimental_enable_bitpacked_activations=True,
-        )
+        super().convert()
 
         model_ir = XCOREModel.deserialize(self._model)
         pass_mgr = PassManager(model_ir)
@@ -184,68 +182,3 @@ class LarqConverter(KerasModelConverter):
         model_ir.buffers = [b] + model_ir.buffers
 
         self._model = model_ir.serialize()
-
-
-#  ----------------------------------------------------------------------------
-#                                   RUNNERS
-#  ----------------------------------------------------------------------------
-
-
-class BinarizedTestRunner(IntegrationTestRunner):
-    _model_generator: LarqCompositeTestModelGenerator
-
-    class OutputData(NamedTuple):
-        reference_quant: np.ndarray
-        xcore: np.ndarray
-
-    def __init__(
-        self,
-        generator: Type[LarqCompositeTestModelGenerator],
-        *,
-        use_device: bool = False,
-    ) -> None:
-        super().__init__(generator, use_device=use_device)
-
-        self._lce_converter = self.make_lce_converter()
-        self.register_converter(self._lce_converter)
-
-        self._lce_evaluator = LarqEvaluator(
-            self, self.get_representative_data, self._lce_converter.get_converted_model
-        )
-        self.register_evaluator(self._lce_evaluator)
-
-    def get_xcore_evaluation_data(self) -> Union[np.ndarray, tf.Tensor]:
-        return self.get_representative_data()
-
-    def make_lce_converter(self) -> LarqConverter:
-        return LarqConverter(self, self.get_built_model)
-
-    def make_repr_data_factory(self) -> InputInitializerDataFactory:
-        return InputInitializerDataFactory(
-            self,
-            lambda: get_bitpacked_shape(self._model_generator.input_shape),
-            dtype=tf.int32,
-        )
-
-    def get_xcore_reference_model(self) -> TFLiteModel:
-        return self._lce_converter.get_converted_model()
-
-    def run(self) -> None:
-        super().run()
-        self._lce_converter.convert()
-        self._lce_evaluator.evaluate()
-
-        self.rerun_post_cache()
-
-    def rerun_post_cache(self) -> None:
-        super().rerun_post_cache()
-
-        self.outputs = self.OutputData(
-            self._lce_evaluator.output_data, self._xcore_evaluator.output_data,
-        )
-        self.converted_models.update(
-            {
-                "reference_lce": self._lce_converter._model,
-                "xcore": self._xcore_converter._model,
-            }
-        )
