+++ test/model_generation/integration_test/test_single_op_models/test_binarized/test_bsign.py
@@ -11,7 +11,7 @@ from tflite2xcore.model_generation.data_factories import InputInitializerDataFac
 from . import (
     BinarizedTestRunner,
     LarqCompositeTestModelGenerator,
-    LarqConverter,
+    LarqSingleOpConverter,
 )
 
 from . import (  # pylint: disable=unused-import
@@ -43,8 +43,8 @@ GENERATOR = BSignTestModelGenerator
 
 
 class BSignTestRunner(BinarizedTestRunner):
-    def make_lce_converter(self) -> LarqConverter:
-        return LarqConverter(self, self.get_built_model, remove_last_op=True)
+    def make_lce_converter(self) -> LarqSingleOpConverter:
+        return LarqSingleOpConverter(self, self.get_built_model, remove_last_op=True)
 
     def _set_config(self, cfg: Configuration) -> None:
         cfg["input_range"] = cfg.pop(
